Migração do PostgresSQL para o MySQL

+++ Api/Enums/EarningType.cs
@@ -1,6 +1,6 @@
 namespace Cashflow.Api.Enums
 {
-    public enum EarningType : short
+    public enum EarningType : byte
     {
         MonthyBenefit = 1,
